+++ src/http/proxy/format/templatize.js
@@ -2,7 +2,7 @@ module.exports = function templatizeResponse (params) {
   let { isBinary, assets, response, isLocal=false } = params
 
   // Bail early
-  if (isBinary || !assets) {
+  if (isBinary) {
     return response
   }
   else {
@@ -14,8 +14,12 @@ module.exports = function templatizeResponse (params) {
     response.body = body.replace(staticRegex, function fingerprint(match) {
       let start = match.startsWith(`\${STATIC(`) ? 10 : 14
       let Key = match.slice(start, match.length-3)
-      if (assets[Key] && !isLocal) {
-        Key = assets[Key]
+      // Normalize around no leading slash for static manifest lookups
+      let startsWithSlash = Key.startsWith('/')
+      let lookup = startsWithSlash ? Key.substr(1) : Key
+      if (assets && assets[lookup] && !isLocal) {
+        Key = assets[lookup]
+        Key = startsWithSlash ? `/${Key}` : Key
       }
       return Key
     })