+++ changelog.md
@@ -4,7 +4,23 @@
 Also see: [Architect changelog](https://github.com/architect/architect/blob/master/changelog.md)
 ---
 
-## [3.10.0] 2020-05-17
+## [3.11.0] 2020-06-07
+
+### Added
+
+- Adds fingerprinting support for pretty URLs and custom 404s in proxy
+- Adds support for `ARC_STATIC_PREFIX` env var to provide symmetry with the new `@static prefix` setting
+  - The `ARC_STATIC_FOLDER` env var is now considered deprecated, and will be removed in a future (breaking) release
+- Added support for leading slashes in build-free templating (e.g. `${arc.static('/this-is-fine.gif')}`)
+
+
+### Fixed
+
+- Ensures the build-free templating engine always runs, not just when fingerprint is enabled
+
+---
+
+## [3.10.0 - 3.10.1] 2020-05-17
 
 ### Added